Adds Many Things (Please Read Desc)
Due to a major complication with git, caching, history, etc., I had to submit a vast amounts of commits into one single commit. If you are reading this, it was a success and I highly not recommend doing what I did. It was either this large commit or a reset of the repo.